Start work on migration hardcoded GUI to scripts.
Changed script stuff:
- added finish handlers;
- added message_box handler;
- added CustomCall;
- added CloneObject;
- added HandleHardcodedScreenMouse, HandleHardcodedScreenKey;
- modified GetTextInfo;
- modified ShowScreen, HideScreen;
- modified MouseClick;
- added bool __DebugInfo;
- added int __SoundVolume;
- added int __MusicVolume.
- fixed GetMonitorHex;
- modified DrawSprite.

Reworked version indexing.

git-svn-id: https://cvet.by:8443/svn/FOnlineSource@632 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: dd52def5884e7a0f2045d4249fd02f3a18a9a4cb

+++ Source/CritterCl.cpp
@@ -688,7 +688,7 @@ void CritterCl::DrawStay( Rect r )
         if( anim )
         {
             uint spr_id = ( IsLife() ? anim->Ind[ 0 ] : anim->Ind[ anim->CntFrm - 1 ] );
-            SprMngr.DrawSpriteSize( spr_id, r.L, r.T, (float) r.W(), (float) r.H(), false, true );
+            SprMngr.DrawSpriteSize( spr_id, r.L, r.T, r.W(), r.H(), false, true );
         }
     }
     else if( Anim3dStay )
